Cleanup some old test files
